Consolidate with other parser tests
